chore: sinon restore stub’ed methods
Setting an argument property was triggering an eslint error. It’s a common ava pattern; overriding the eslint rule in that case would be acceptable.

But since Sinon can take of this task, we might as well let it do it.

Note that I make sure to always run the `afterEach` routine so a failing task doesn’t break following tests.